Set default background color to white.
Since we are setting the text color of the body, we should set the
background color as well. While most browsers default to white for the
background, we shouldn't assume that the default background color will
be white.

For example, in Firefox you can change the default text and background
color: https://support.mozilla.org/en-US/kb/change-fonts-and-colors-websites-use